Active record struggle

+++ db/migrate/20210105014203_create_items.rb
@@ -3,10 +3,9 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.float :unit_price
+      t.integer :unit_price
 
       t.timestamps
     end
   end
 end
-
